Move most Scarpe-components stuff into a Components namespace
