+++ lib/scarpe/wv.rb
@@ -8,8 +8,8 @@ require "securerandom"
 require "json"
 
 require "bloops"
-require "scarpe/modular_logger"
-require "scarpe/promises"
+require "scarpe/components/modular_logger"
+require "scarpe/components/promises"
 
 # Set up hierarchical logging using the SCARPE_LOG_CONFIG var for configuration
 log_config = if ENV["SCARPE_LOG_CONFIG"]
@@ -18,7 +18,7 @@ else
   ENV["SCARPE_DEBUG"] ? Shoes::Log::DEFAULT_DEBUG_LOG_CONFIG : Shoes::Log::DEFAULT_LOG_CONFIG
 end
 
-Shoes::Log.instance = Scarpe::ModularLogImpl.new
+Shoes::Log.instance = Scarpe::Components::ModularLogImpl.new
 Shoes::Log.configure_logger(log_config)
 
 require_relative "wv/web_wrangler"
