+++ lib/scarpe/wv/image.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-require "scarpe/base64"
+require "scarpe/components/base64"
 
 class Scarpe
   class WebviewImage < WebviewWidget
-    include Base64
+    include Components::Base64
     def initialize(properties)
       super
