+++ scarpe-components/lib/scarpe/components/unit_test_helpers.rb
