+++ test/test_helper.rb
@@ -5,7 +5,7 @@ RUBY_MAIN_OBJ = self
 $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
 require "scarpe"
 require "scarpe/evented_assertions"
-require "scarpe/unit_test_helpers"
+require "scarpe/components/unit_test_helpers"
 
 require "json"
 require "fileutils"
